mw-dapp/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-parler==2.3
django-parler-rest==2.2
django-resized==1.0.2
djangorestframework==3.15.2
pillow==10.3.0
sqlparse==0.5.0
